+++ CHANGELOG.md
@@ -7,6 +7,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+
+## [2.0.0-beta.3] - 2023-07-09
+
 ### Added
 
 * The `Amount` type is now exhaustive
@@ -434,7 +437,8 @@ Announce new status of the project in readme: The project is now in "maintenance
  - `Parser` iterator type
 
 
-[Unreleased]: https://github.com/jcornaz/beancount-parser/compare/v2.0.0-beta.2...HEAD
+[Unreleased]: https://github.com/jcornaz/beancount-parser/compare/v2.0.0-beta.3...HEAD
+[2.0.0-beta.3]: https://github.com/jcornaz/beancount-parser/compare/v2.0.0-beta.2...v2.0.0-beta.3
 [2.0.0-beta.2]: https://github.com/jcornaz/beancount-parser/compare/v2.0.0-beta.1...v2.0.0-beta.2
 [2.0.0-beta.1]: https://github.com/jcornaz/beancount-parser/compare/v1...v2.0.0-beta.1
 [1.16.2]: https://github.com/jcornaz/beancount-parser/compare/v1.16.1...v1.16.2